Fix type

+++ util/oceanutil.py
@@ -233,7 +233,7 @@ def createFREFromDatatoken(
 
 
 @enforce_types
-def set_allocation(amount: float, nft_addr: str, chainID: int, from_account):
+def set_allocation(amount: int, nft_addr: str, chainID: int, from_account):
     veAllocate().setAllocation(amount, nft_addr, chainID, {"from": from_account})
